+++ packages/provider/src/image-model/v3/image-model-v3-call-options.ts
