+++ packages/provider/src/image-model/v3/index.ts
@@ -1,6 +1,6 @@
 export type {
   ImageModelV2,
   ImageModelV2ProviderMetadata,
-} from './image-model-v2';
-export type { ImageModelV2CallOptions } from './image-model-v2-call-options';
-export type { ImageModelV2CallWarning } from './image-model-v2-call-warning';
+} from './image-model-v3';
+export type { ImageModelV2CallOptions } from './image-model-v3-call-options';
+export type { ImageModelV2CallWarning } from './image-model-v3-call-warning';
