+++ packages/provider/src/provider/v2/provider-v2.ts
@@ -1,5 +1,5 @@
 import { EmbeddingModelV2 } from '../../embedding-model/v2/embedding-model-v2';
-import { ImageModelV2 } from '../../image-model/v3/image-model-v2';
+import { ImageModelV2 } from '../../image-model/v3/image-model-v3';
 import { LanguageModelV2 } from '../../language-model/v2/language-model-v2';
 import { SpeechModelV2 } from '../../speech-model/v2/speech-model-v2';
 import { TranscriptionModelV2 } from '../../transcription-model/v2/transcription-model-v2';
